consolidate attestation signing into attestor/evm and proofgen/attestation, and centralize relay types in types/v2 to break the proofgen/txbuilder import cycle
