--- link/internal/attestation/evm/signing_test.go
@@ -1,168 +1,0 @@
-package evm
-
-import (
-	"context"
-	"crypto/sha256"
-	"testing"
-
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func TestDigestMatchesFormula(t *testing.T) {
-	data := []byte("some attestation data")
-
-	for _, tag := range []byte{TagStateAttestation, TagPacketAttestation} {
-		got := Digest(tag, data)
-
-		inner := sha256.Sum256(data)
-		want := sha256.Sum256(append([]byte{tag}, inner[:]...))
-
-		assert.Equal(t, want, got, "tag %x", tag)
-	}
-}
-
-type stubSigner struct {
-	sig []byte
-	err error
-}
-
-func (s stubSigner) Sign(context.Context, []byte) ([]byte, error) {
-	return s.sig, s.err
-}
-
-func TestSignABI(t *testing.T) {
-	t.Run("signsDigestAndNormalizesRecoveryID", func(t *testing.T) {
-		sig := make([]byte, 65)
-		sig[64] = 1 // raw 0/1 recovery id
-
-		signer := stubSigner{sig: sig}
-
-		got, err := SignABI(context.Background(), signer, TagStateAttestation, []byte("data"))
-		require.NoError(t, err)
-		require.Len(t, got, 65)
-		assert.Equal(t, byte(28), got[64], "recovery id must be normalized to Solidity's 27/28 form")
-	})
-
-	t.Run("propagatesSignerError", func(t *testing.T) {
-		signer := stubSigner{err: assert.AnError}
-
-		_, err := SignABI(context.Background(), signer, TagStateAttestation, []byte("data"))
-		require.ErrorIs(t, err, assert.AnError)
-	})
-}
-
-func TestNormalizeSignature(t *testing.T) {
-	for _, tt := range []struct {
-		name        string
-		recoveryID  byte
-		expected    byte
-		errContains string
-	}{
-		{name: "raw0", recoveryID: 0, expected: 27},
-		{name: "raw1", recoveryID: 1, expected: 28},
-		{name: "legacy27", recoveryID: 27, expected: 27},
-		{name: "legacy28", recoveryID: 28, expected: 28},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			sig := make([]byte, 65)
-			sig[64] = tt.recoveryID
-
-			got, err := normalizeSignature(sig)
-			require.NoError(t, err)
-			assert.Equal(t, tt.expected, got[64])
-		})
-	}
-
-	t.Run("rejectsInvalidRecoveryID", func(t *testing.T) {
-		sig := make([]byte, 65)
-		sig[64] = 99
-
-		_, err := normalizeSignature(sig)
-		require.ErrorContains(t, err, "invalid recovery ID")
-	})
-
-	t.Run("rejectsWrongLength", func(t *testing.T) {
-		_, err := normalizeSignature([]byte{0x01})
-		require.ErrorContains(t, err, "invalid signature length")
-	})
-
-	t.Run("doesNotMutateInput", func(t *testing.T) {
-		sig := make([]byte, 65)
-		sig[64] = 0
-
-		_, err := normalizeSignature(sig)
-		require.NoError(t, err)
-		assert.Equal(t, byte(0), sig[64], "input must not be mutated")
-	})
-}
-
-func TestRecoverSigner(t *testing.T) {
-	t.Run("roundTrip", func(t *testing.T) {
-		key, err := crypto.GenerateKey()
-		require.NoError(t, err)
-
-		expected := crypto.PubkeyToAddress(key.PublicKey)
-
-		digest := Digest(TagPacketAttestation, []byte("attested data"))
-
-		sig, err := crypto.Sign(digest[:], key)
-		require.NoError(t, err)
-		require.Len(t, sig, 65)
-
-		recovered, err := RecoverSigner(digest, sig)
-		require.NoError(t, err)
-		require.Equal(t, expected, recovered)
-	})
-
-	t.Run("rejectsWrongLength", func(t *testing.T) {
-		_, err := RecoverSigner([32]byte{}, []byte{0x01, 0x02})
-		require.Error(t, err)
-	})
-
-	// AcceptsLegacyVByte verifies RecoverSigner works against signatures
-	// using Ethereum's legacy v (27/28), the convention the real attestor
-	// signs with (see SignABI/normalizeSignature), not just the raw 0/1
-	// recovery id crypto.Sign itself produces.
-	t.Run("acceptsLegacyVByte", func(t *testing.T) {
-		key, err := crypto.GenerateKey()
-		require.NoError(t, err)
-
-		expected := crypto.PubkeyToAddress(key.PublicKey)
-
-		digest := Digest(TagPacketAttestation, []byte("attested data"))
-
-		sig, err := crypto.Sign(digest[:], key)
-		require.NoError(t, err)
-
-		legacySig := make([]byte, 65)
-		copy(legacySig, sig)
-		legacySig[64] += 27
-
-		recovered, err := RecoverSigner(digest, legacySig)
-		require.NoError(t, err)
-		require.Equal(t, expected, recovered)
-	})
-
-	t.Run("interopsWithSignABI", func(t *testing.T) {
-		key, err := crypto.GenerateKey()
-		require.NoError(t, err)
-
-		expected := crypto.PubkeyToAddress(key.PublicKey)
-
-		digest := Digest(TagStateAttestation, []byte("data"))
-
-		rawSig, err := crypto.Sign(digest[:], key)
-		require.NoError(t, err)
-
-		signer := stubSigner{sig: rawSig}
-
-		sig, err := SignABI(context.Background(), signer, TagStateAttestation, []byte("data"))
-		require.NoError(t, err)
-
-		recovered, err := RecoverSigner(digest, sig)
-		require.NoError(t, err)
-		require.Equal(t, expected, recovered, "RecoverSigner must accept exactly what SignABI produces")
-	})
-}
