+++ link/internal/relay/txbuilder/txbuilder.go
@@ -2,49 +2,24 @@
 // update details. There is one implementation per supported chain type.
 package txbuilder
 
-import channeltypesv2 "github.com/cosmos/ibc-go/v11/modules/core/04-channel/v2/types"
+import (
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/pkg/errors"
 
-// RelayKind the packet operation one PacketRelayItem asks to perform.
-type RelayKind int
-
-// Relay kinds
-const (
-	KindUnknown RelayKind = iota
-	KindRecv
-	KindAck
-	KindTimeout
+	"github.com/cosmos/ibc/link/internal/config"
+	"github.com/cosmos/ibc/link/internal/relay/txbuilder/evm"
+	v2 "github.com/cosmos/ibc/link/internal/types/v2"
 )
-
-// PacketRelayItem one packet operation to include in a relay tx, along with
-// the membership/non-membership proof authorizing it.
-type PacketRelayItem struct {
-	Kind        RelayKind
-	Packet      channeltypesv2.Packet
-	Acks        [][]byte // one per payload, order preserved; only set for KindAck
-	Proof       []byte
-	ProofHeight uint64
-}
-
-// ClientUpdate the state proof to update a destination client with before
-// any packet operations in the same tx are processed.
-type ClientUpdate struct {
-	ClientID   string
-	StateProof []byte
-}
-
-// RelayTx one transaction ready to submit, targeting To with calldata Data.
-type RelayTx struct {
-	To   []byte
-	Data []byte
-}
 
 // TxBuilder builds transactions from packet relay and client update details
 // for one chain. It returns a list of transactions rather than one, since
 // some chains (e.g. Solana, due to tx size limits) must split a batch of
 // packet relays across multiple transactions.
 type TxBuilder interface {
-	BuildRelayTxs(clientUpdate ClientUpdate, packetRelayItems []PacketRelayItem) ([]RelayTx, error)
+	BuildRelayTxs(clientUpdate v2.ClientUpdate, packetRelayItems []v2.PacketRelayItem) ([]v2.RelayTx, error)
 }
+
+var _ TxBuilder = (*evm.TxBuilder)(nil)
 
 // Set resolves a TxBuilder by chain id.
 type Set struct {
@@ -62,4 +37,25 @@ func NewSet(builders map[string]TxBuilder) *Set {
 func (s *Set) Get(chainID string) (TxBuilder, bool) {
 	builder, ok := s.builders[chainID]
 	return builder, ok
+}
+
+// NewSetFromConfig builds a Set with one EVM TxBuilder per configured EVM
+// chain, bound to that chain's configured ICS26Router address.
+func NewSetFromConfig(cfg config.Config) (*Set, error) {
+	builders := make(map[string]TxBuilder, len(cfg.Chains))
+
+	for _, chainCfg := range cfg.Chains {
+		switch chainCfg.Type() {
+		case config.ChainTypeEVM:
+			if !common.IsHexAddress(chainCfg.EVM.ICS26Router) {
+				return nil, errors.Errorf("chain %q: invalid ics26 router address %q", chainCfg.ChainID, chainCfg.EVM.ICS26Router)
+			}
+
+			builders[chainCfg.ChainID] = evm.New(common.HexToAddress(chainCfg.EVM.ICS26Router))
+		default:
+			return nil, errors.Errorf("chain %q: unsupported chain type %q for tx building", chainCfg.ChainID, chainCfg.Type())
+		}
+	}
+
+	return NewSet(builders), nil
 }
