+++ link/internal/service/attestor/local.go
@@ -14,7 +14,7 @@ import (
 
 	channeltypesv2 "github.com/cosmos/ibc-go/v11/modules/core/04-channel/v2/types"
 	hostv2 "github.com/cosmos/ibc-go/v11/modules/core/24-host/v2"
-	evm "github.com/cosmos/ibc/link/internal/attestation/evm"
+	evm "github.com/cosmos/ibc/link/internal/service/attestor/evm"
 	v2 "github.com/cosmos/ibc/link/internal/types/v2"
 )
