+++ link/internal/types/v2/types.go
@@ -27,6 +27,51 @@ const (
 	WriteAckStatusError
 )
 
+// ProofKind the kind of packet claim a proof attests to.
+type ProofKind int
+
+// Proof kinds
+const (
+	ProofKindUnknown ProofKind = iota
+	ProofKindPacketCommitment
+	ProofKindAcknowledgement
+	ProofKindReceiptAbsence
+)
+
+// RelayKind the packet operation one PacketRelayItem asks to perform.
+type RelayKind int
+
+// Relay kinds
+const (
+	RelayKindUnknown RelayKind = iota
+	RelayKindRecv
+	RelayKindAck
+	RelayKindTimeout
+)
+
+// PacketRelayItem one packet operation to include in a relay tx, along with
+// the membership/non-membership proof authorizing it.
+type PacketRelayItem struct {
+	Kind        RelayKind
+	Packet      channeltypesv2.Packet
+	Acks        [][]byte // one per payload, order preserved; only set for RelayKindAck
+	Proof       []byte
+	ProofHeight uint64
+}
+
+// ClientUpdate the state proof to update a destination client with before
+// any packet operations in the same tx are processed.
+type ClientUpdate struct {
+	ClientID   string
+	StateProof []byte
+}
+
+// RelayTx one transaction ready to submit, targeting To with calldata Data.
+type RelayTx struct {
+	To   []byte
+	Data []byte
+}
+
 // PacketEvent a packet event.
 type PacketEvent struct {
 	Height    uint64
